rename package

+++ src/database/migrations/2012_01_01_000120_market_create_table_product.php
@@ -10,22 +10,25 @@ class MarketCreateTableProduct extends Migration
 	 * @return void
 	 */
 	public function up()
-	{   
-            Schema::create('012_111_product', function($table) {
-                $table->engine = 'InnoDB';
-                $table->increments('id_111')->unsigned();
+	{
+		if (!Schema::hasTable('012_111_product'))
+		{
+			Schema::create('012_111_product', function ($table) {
+				$table->engine = 'InnoDB';
+				$table->increments('id_111')->unsigned();
 				$table->integer('custom_field_group_111')->unsigned()->nullable();
 				$table->tinyInteger('price_type_111')->unsigned();
 				$table->decimal('price_111', 10, 2)->nullable();
 				$table->decimal('weight_111', 10, 3)->nullable();
 				$table->boolean('active_111');
 
-				$table->string('data_lang_111',255)->nullable();
+				$table->string('data_lang_111', 255)->nullable();
 				$table->text('data_111')->nullable();
 
 				$table->foreign('custom_field_group_111', 'fk01_012_111_product')->references('id_025')->on('001_025_field_group')
-					->onDelete('restrict')->onUpdate('cascade');
-            });
+						->onDelete('restrict')->onUpdate('cascade');
+			});
+		}
 	}
 
 	/**
@@ -35,7 +38,10 @@ class MarketCreateTableProduct extends Migration
 	 */
 	public function down()
 	{
-            Schema::drop('012_111_product');
+		if (Schema::hasTable('012_111_product'))
+		{
+			Schema::drop('012_111_product');
+		}
 	}
 
 }
